Create the sfo role and schema on launch
I've had to use a more 'sequelize' format for the migrations' names as
they share a 'namespace' with the other apps - such as TR that has
already migrated. We should probably go back and amend TR and the DB log
to match this new style for consistency's sake.

Issue: Scottish-Natural-Heritage/Licensing#287