Remove the DB password assertions
To be able to run the pre-migrations, migrations and let the API connect
to the database we need to supply two different passwords for the
Postgres connections to be happy. Sqlite doesn't care about passwords,
so we don't **need** these passwords for a local dev & test environment.

Issue: Scottish-Natural-Heritage/Licensing#395